requirements.txt
--find-links https://download.pytorch.org/whl/cpu
fastapi[all]
jinja2
transformers
torch>=2.0+cpu
Pillow